+++ src/main/java/frc/simulacion/elmejorgrupo/tpcinco/generadores/Generador.java
@@ -3,7 +3,7 @@ package frc.simulacion.elmejorgrupo.tpcinco.generadores;
 import frc.simulacion.elmejorgrupo.tpcinco.model.TiposAuto;
 
 public class Generador {
-    public static TiposAuto generarAuto(Float RND){
+    public static TiposAuto generarAuto(Double RND){
         if (RND < 0.45f){
             return TiposAuto.PEQUENIO;
         } else if (RND < 0.70f) {
@@ -13,20 +13,20 @@ public class Generador {
         }
     }
 
-    public static Float generarTiempoLlegadaAuto(Float RND){
+    public static Double generarTiempoLlegadaAuto(Double RND){
         //RND/=10; esto fuerza que aparezcan autos sin atender
-        return (float) (-13f * Math.log(1-RND));
+        return (Double) (-13f * Math.log(1-RND));
     }
 
-    public static float generarTiempoEstacionamiento(Float RND){
+    public static Double generarTiempoEstacionamiento(Double RND){
         if (RND < 0.5f){
-            return 60f;
+            return (double) 60f;
         } else if (RND < 0.8f) {
-            return 120f;
+            return (double) 120f;
         } else if (RND < 0.95){
-            return 180f;
+            return (double) 180f;
         } else {
-            return 240f;
+            return (double) 240f;
         }
     }
 }
